+++ packages/valibot/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @ai-sdk/valibot
 
+## 3.0.0-beta.14
+
+### Patch Changes
+
+- @ai-sdk/provider-utils@5.0.0-beta.14
+
+## 3.0.0-beta.13
+
+### Patch Changes
+
+- Updated dependencies [b0c2869]
+- Updated dependencies [7e26e81]
+  - @ai-sdk/provider-utils@5.0.0-beta.13
+
+## 3.0.0-beta.12
+
+### Patch Changes
+
+- Updated dependencies [46d1149]
+  - @ai-sdk/provider-utils@5.0.0-beta.12
+
 ## 3.0.0-beta.11
 
 ### Patch Changes